bcrypt and jwt auth implemented

+++ src/controllers/contactController.js
@@ -1,14 +1,16 @@
 import * as contactService from "../service/contactService.js";
 import { responseHandler } from "../commonFunction/responsehandler.js";
+import errHandler from "../middleware/errHandler.js";
 
 //@desc Get all contacts
 //@route GET /api/getContacts
-//@access public 
+//@access private 
 const getAllContactscontroller = async (req, res) => {
 
     try {
-        const data = await contactService.getAllUsersService();
-        responseHandler(res, 200, `data fetched successfully`, data);
+        const created_by = req.user.email;
+        const data = await contactService.getAllUsersService(created_by);
+        data.length === 0 ? responseHandler(res, 200, `No data found`, data) : responseHandler(res, 200, `data fetched successfully`, data);
     } catch (error) {
         next(error);
     }
@@ -17,7 +19,7 @@ const getAllContactscontroller = async (req, res) => {
 
 //@desc Get contact by ID
 //@route GET /api/getContact/:id
-//@access public
+//@access private
 const getContactByIdcontroller = async (req, res) => {
 
     try {
@@ -34,46 +36,62 @@ const getContactByIdcontroller = async (req, res) => {
 
 //@desc Create new contact
 //@route POST /api/createContact
-//@access public
+//@access private
 const createContactcontroller = async (req, res) => {
-try {
-     const { name, email, password } = req.body;
-    if (!name || !email || !password) {
-        res.status(400)
-        throw new Error("All fields are mandatory!");
+    try {
+        const { name, email, phone } = req.body;
+        const created_by = req.user.email;
+        if (!name || !email || !phone) {
+            res.status(400)
+            throw new Error("All fields are mandatory!");
+        }
+        const data = await contactService.createUserService(name, email, phone, created_by);
+        responseHandler(res, 201, `New contact created`, data)
+    } catch (error) {
+        next(error);
     }
-    const data = await contactService.createUserService(name, email, password);
-    responseHandler(res, 200, `New contact created`, data)
-} catch (error) {
-    next(error);
-}
-   
+
 };
 
 //@desc Update contact by ID
 //@route PUT /api/updateContact/:id
-//@access public
+//@access private
 const updateContactcontroller = async (req, res) => {
-   try {
-       const { name, email } = req.body;
-       if (!name || !email) {
-           res.status(400)
-           throw new Error("All fields are mandatory!");
-       }
-       const data = await contactService.updateUserService(req.params.id, name, email);
-       responseHandler(res, 200, `Contact updated in ${req.params.id}`, data);
-   } catch (error) {
-       next(error);
-   }
+    try {
+        console.log('inside update controller');
+        
+        const { name, email } = req.body;
+        const created_by = req.user.email;
+        if (!name || !email) {
+            res.status(400)
+            throw new Error("All fields are mandatory!");
+        }
+        const validUser = await contactService.getUsersbyIdService(req.params.id);
+
+        if (validUser[0].created_by !== created_by) {
+            res.status(403)
+            throw new Error("You are not authorized to update this contact");
+        }
+        const data = await contactService.updateUserService(req.params.id, name, email);
+        responseHandler(res, 200, `Contact updated in ${req.params.id}`, data);
+    } catch (error) {
+        errHandler(error, req, res);}
 
 };
 
 //@desc Delete contact by ID
 //@route DELETE /api/deleteContact/:id
-//@access public
+//@access private
 const deleteContactcontroller = async (req, res) => {
 
     try {
+        const created_by = req.user.email;
+        const validUser = await contactService.getUsersbyIdService(req.params.id);
+
+        if (validUser[0].created_by !== created_by) {
+            res.status(403)
+            throw new Error("You are not authorized to delete this contact");
+        }
         const data = await contactService.deleteUserByIdService(req.params.id);
         responseHandler(res, 200, `Contact deleted in ${req.params.id}`, data);
     } catch (error) {
